✅ Implemented notes with somewhat authentication and authorization

+++ src/features/note/services/note.service.ts
@@ -1,16 +1,34 @@
 import { injectable, singleton } from 'tsyringe';
 import NoteRepository from '../repository/note.reposotory';
-import { NoteDto, NoteFilter } from '../types';
+import { NoteFilter } from '../types';
+import { CreateNoteDTO, UpdateNoteDTO } from '../dto';
 import { logger } from '../../../shared/utils/logger';
+import { Note } from '../models/note.model';
+import { User } from '@/features/user/models/user.model';
+import { ServiceStatus } from '@/shared/utils/constants';
+import { StatusCodes } from 'http-status-codes';
+import ServiceException from '@/shared/utils/serverException';
 
 @injectable()
 @singleton()
 export default class NoteService {
   constructor(private noteRepository: NoteRepository) {}
 
-  async createNote(data: NoteDto) {
+  private ensureOwnership(note: Note, userId: number) {
+    if (note.userId !== userId) {
+      throw new Error('Unauthorized: You do not own this note');
+    }
+  }
+
+  async createNote(data: CreateNoteDTO, userId: number) {
     try {
-      const note = await this.noteRepository.createNote(data);
+      const note = await this.noteRepository.createNote({
+        title: data.title,
+        content: data.content,
+        isPrivate: data.isPrivate,
+        user: { id: userId } as User,
+      });
+
       return {
         message: 'Note created successfully.',
         noteTitle: note.title,
@@ -21,9 +39,13 @@ export default class NoteService {
     }
   }
 
-  async updateNote(id: number, data: NoteDto) {
+  async updateNote(id: number, userId: number, data: UpdateNoteDTO) {
     try {
-      const note = await this.noteRepository.updateNote(id, data);
+      let note = await this.noteRepository.findNoteById(id);
+      if (note) {
+        this.ensureOwnership(note, userId);
+      }
+      note = await this.noteRepository.updateNote(id, data);
       return {
         message: 'Note updated successfully.',
         noteTitle: note?.title,
@@ -34,11 +56,18 @@ export default class NoteService {
     }
   }
 
-  async deleteNote(id: number) {
+  async deleteNote(id: number, userId: number) {
     try {
-      const note = await this.noteRepository.deleteNote(id);
+      let note = await this.noteRepository.findNoteById(id);
+      if (!note) {
+        throw new ServiceException(StatusCodes.NOT_FOUND, ServiceStatus.FAILURE, 'Note not found');
+      }
+      this.ensureOwnership(note, userId);
+
+      const noteTitle = await this.noteRepository.deleteNote(id);
+
       return {
-        message: `${note} deleted successfully`,
+        message: ` ${noteTitle} deleted successfully`,
       };
     } catch (error: any) {
       logger.error('Error while deleting note:', error);
@@ -46,27 +75,44 @@ export default class NoteService {
     }
   }
 
-  async getNoteById(id: number) {
+  async getNoteById(id: number, userId: number) {
     try {
-      return await this.noteRepository.findNoteById(id);
+      const note = await this.noteRepository.findPublicNoteById(id);
+      if (note) {
+        this.ensureOwnership(note, userId);
+      }
+      return note;
     } catch (error: any) {
       logger.error('Error while deleting note:', error);
       throw error;
     }
   }
 
-  async getNotes(filters: NoteFilter) {
+  async getNotes(filters: NoteFilter, userId: number) {
     try {
-      return await this.noteRepository.getNotes(filters);
+      return await this.noteRepository.getNotes(filters, userId);
     } catch (error: any) {
       logger.error('Error in getting notes:', error);
       throw error;
     }
   }
 
-  async getPrivateNotes(filters: NoteFilter) {
+  async getPrivateNotes(filters: NoteFilter, userId: number) {
     try {
-      return await this.noteRepository.getPrivateNotes(filters);
+      return await this.noteRepository.getPrivateNotes(filters, userId);
+    } catch (error: any) {
+      logger.error('Error while getting private notes:', error);
+      throw error;
+    }
+  }
+  async getPrivateNoteById(id: number, userId: number) {
+    try {
+      const note = await this.noteRepository.findPrivateNoteById(id);
+      if (!note) {
+        throw new ServiceException(StatusCodes.NOT_FOUND, ServiceStatus.FAILURE, 'Note not found');
+      }
+      this.ensureOwnership(note, userId);
+      return note;
     } catch (error: any) {
       logger.error('Error while getting private notes:', error);
       throw error;
